LU-4669 ldiskfs: do not export static inline functions
In ldiskfs/ldiskfs_extents.h, both ldiskfs_ext_store_pblock()
and ldiskfs_ext_pblock() are defined as static inline functions.
However, commit 49b06fba39e (http://review.whamcloud.com/4804)
exported these two functions in ldiskfs/extents.c, which is not
allowed by gcc and causes the following compile errors:

error: __ksymtab_ldiskfs_ext_pblock causes a section type conflict
error: __ksymtab_ldiskfs_ext_store_pblock causes a section type conflict

This patch fixes the above errors by not exporting the two functions.

The patch also fixes the following compile error by removing the
dead codes: "error: 'lmv_adjust_dirpages' defined but not used".
This code was moved to mdc_adjust_dirpages() in commit 4e0c8aeb94
(http://review.whamcloud.com/7043).

And also fixes the following compile errors by using "LPX64”:
error: format '%llx' expects type 'long long unsigned int',
       but argument 5 has type '__u64'
error: format '%#llx' expects type 'long long unsigned int',
       but argument 3 has type '__u64'

+++ lustre/include/darwin/lustre_lite.h
@@ -79,7 +79,7 @@ struct lookup_intent {
 	int     it_magic;
 	void    (*it_op_release)(struct lookup_intent *);
 	int     it_op;
-	int     it_flags;
+	__u64   it_flags;
 	int     it_create_mode;
 	union {
                 struct lustre_intent_data lustre;
